+++ src/main/java/me/towdium/jecalculation/data/label/labels/LPlaceholder.java
@@ -39,11 +39,11 @@ public class LPlaceholder extends ILabel.Impl {
         name = tag.getString(KEY_NAME);
     }
 
-    public LPlaceholder(String name, int amount) {
+    public LPlaceholder(String name, long amount) {
         this(name, amount, false);
     }
 
-    public LPlaceholder(String name, int amount, boolean silent) {
+    public LPlaceholder(String name, long amount, boolean silent) {
         super(amount, false);
         this.name = name;
         if (!silent) getRecord().push(new LPlaceholder(name, 1, true));
@@ -88,7 +88,7 @@ public class LPlaceholder extends ILabel.Impl {
 
     @Override
     public int hashCode() {
-        return name.hashCode() ^ amount;
+        return name.hashCode() ^ (int) amount;
     }
 
     @Nullable
